// autogpt_platform/frontend/src/app/api/__generated__/models/oAuth2Credentials.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { OAuth2CredentialsTitle } from "./oAuth2CredentialsTitle";
import type { OAuth2CredentialsUsername } from "./oAuth2CredentialsUsername";
import type { OAuth2CredentialsAccessTokenExpiresAt } from "./oAuth2CredentialsAccessTokenExpiresAt";
import type { OAuth2CredentialsRefreshToken } from "./oAuth2CredentialsRefreshToken";
import type { OAuth2CredentialsRefreshTokenExpiresAt } from "./oAuth2CredentialsRefreshTokenExpiresAt";
import type { OAuth2CredentialsMetadata } from "./oAuth2CredentialsMetadata";

export interface OAuth2Credentials {
  id?: string;
  provider: string;
  title?: OAuth2CredentialsTitle;
  type?: "oauth2";
  username?: OAuth2CredentialsUsername;
  access_token: string;
  access_token_expires_at?: OAuth2CredentialsAccessTokenExpiresAt;
  refresh_token?: OAuth2CredentialsRefreshToken;
  refresh_token_expires_at?: OAuth2CredentialsRefreshTokenExpiresAt;
  scopes: string[];
  metadata?: OAuth2CredentialsMetadata;
}